Admin can create blog post as user group
- Also created UserGroupHelper to core, practically same method would be
in four(4) helpers otherwise.
