+++ decidim-admin/app/views/decidim/admin/imports/new.html.erb
@@ -19,13 +19,7 @@
       </div>
       <% if current_organization.user_groups_enabled? && Decidim::UserGroups::ManageableUserGroups.for(current_user).verified.any? %>
         <div class="field">
-          <%=
-            form.select(
-              :user_group_id,
-              user_groups.map { |g| [g.name, g.id] },
-              include_blank: current_user.name
-            )
-          %>
+          <%= user_group_select_field form, :user_group_id %>
         </div>
       <% end %>
       <div class="grid-x">
