+++ decidim-blogs/app/forms/decidim/blogs/admin/post_form.rb
@@ -10,8 +10,17 @@ module Decidim
         translatable_attribute :title, String
         translatable_attribute :body, String
 
+        attribute :user_group_id, Integer
+
         validates :title, translatable_presence: true
         validates :body, translatable_presence: true
+
+        def user_group
+          @user_group ||= Decidim::UserGroup.find_by(
+            organization: current_organization,
+            id: user_group_id.to_i
+          )
+        end
       end
     end
   end
