+++ decidim-blogs/app/views/decidim/blogs/admin/posts/_form.html.erb
@@ -3,6 +3,11 @@
     <h2 class="card-title"><%= title %></h2>
   </div>
   <div class="card-section">
+    <% if current_organization.user_groups_enabled? && Decidim::UserGroups::ManageableUserGroups.for(current_user).verified.any? %>
+      <div class="field">
+        <%= user_group_select_field form, :user_group_id %>
+      </div>
+    <% end %>
     <div class="row column">
       <%= form.translated :text_field, :title, autofocus: true %>
     </div>
